1.1.12 - numeric et integer
- Remplacement des required integer par numeric
- Non affichage des thead sur le dashboard si il n y a aucune info

+++ app/Http/Requests/ContractRequest.php
@@ -27,9 +27,9 @@ class ContractRequest extends Request
             'flat' => 'required|integer',
             'tenant' => 'required|integer',
             'start_date' => 'required|date',
-            'price' => 'integer',
-            'application_fee' => 'integer',
-            'deposit' => 'integer'
+            'price' => 'numeric',
+            'application_fee' => 'numeric',
+            'deposit' => 'numeric'
         ];
     }
 }
